// test/fixtures/api.d.ts
export interface paths {
// ...
  "/blogposts/{post_id}": {
      parameters: {
          query?: never;
          header?: never;
          path: {
              post_id: string;
          };
          cookie?: never;
      };
      get: {
          parameters: {
              query?: {
                  version?: number;
                  format?: string;
              };
              header?: never;
              path: {
                  post_id: string;
              };
              cookie?: never;
          };
          requestBody?: never;
          responses: {
              200: components["responses"]["PostGet"];
              404: components["responses"]["Error"];
              500: components["responses"]["Error"];
          };
      };
      put?: never;
      post?: never;
      delete: {
          parameters: {
              query?: never;
              header?: never;
              path: {
                  post_id: string;
              };
              cookie?: never;
          };
          requestBody?: never;
          responses: {
              200: components["responses"]["PostDelete"];
              500: components["responses"]["Error"];
          };
      };
      options?: never;
      head?: never;
      patch: {
          parameters: {
              query?: never;
              header?: never;
              path: {
                  post_id: string;
              };
              cookie?: never;
          };
          requestBody: components["requestBodies"]["PatchPost"];
          responses: {
              200: components["responses"]["PatchPost"];
              404: components["responses"]["Error"];
              500: components["responses"]["Error"];
          };
      };
      trace?: never;
  };
// ...
  "/string-array": {
      parameters: {
          query?: never;
          header?: never;
          path?: never;
          cookie?: never;
      };
      get: {
          parameters: {
              query?: never;
              header?: never;
              path?: never;
              cookie?: never;
          };
          requestBody?: never;
          responses: {
              200: components["responses"]["StringArray"];
              500: components["responses"]["Error"];
          };
      };
      put?: never;
      post?: never;
      delete?: never;
      options?: never;
      head?: never;
      patch?: never;
      trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
      Post: {
          title: string;
          body: string;
          publish_date?: number;
      };
      StringArray: string[];
      User: {
          email: string;
          age?: number;
          avatar?: string;
          /** Format: date */
          created_at: number;
          /** Format: date */
          updated_at: number;
      };
  };
  responses: {
      AllPostsGet: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": components["schemas"]["Post"][];
          };
      };
      CreatePost: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  status: string;
              };
          };
      };
      CreateTag: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  status: string;
              };
          };
      };
      CreateReply: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json;charset=utf-8": {
                  message: string;
              };
          };
      };
      Contact: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "text/html": string;
          };
      };
      EmptyError: {
          headers: {
              [name: string]: unknown;
          };
          content?: never;
      };
      Error: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  code: number;
                  message: string;
              };
          };
      };
      PatchPost: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  status: string;
              };
          };
      };
      PostDelete: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  status: string;
              };
          };
      };
      PostGet: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": components["schemas"]["Post"];
          };
      };
      StringArray: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": components["schemas"]["StringArray"];
          };
      };
      Tag: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": string;
          };
      };
      User: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": components["schemas"]["User"];
          };
      };
      MultipleResponse: {
          headers: {
              [name: string]: unknown;
          };
          content: {
              "application/json": {
                  id: string;
                  email: string;
                  name?: string;
              };
              "application/ld+json": {
                  "@id": string;
                  email: string;
                  name?: string;
              };
          };
      };
  };
  parameters: never;
  requestBodies: {
      CreatePost: {
          content: {
              "application/json": {
                  title: string;
                  body: string;
                  publish_date: number;
              };
          };
      };
      CreatePostOptional: {
          content: {
              "application/json": {
                  title: string;
                  body: string;
                  publish_date: number;
              };
          };
      };
      CreateTag: {
          content: {
              "application/json": {
                  description?: string;
              };
          };
      };
      CreateReply: {
          content: {
              "application/json;charset=utf-8": {
                  message: string;
                  replied_at: number;
              };
          };
      };
      Contact: {
          content: {
              "multipart/form-data": {
                  name: string;
                  email: string;
                  subject: string;
                  message: string;
              };
          };
      };
      PatchPost: {
          content: {
              "application/json": {
                  title?: string;
                  body?: string;
                  publish_date?: number;
              };
          };
      };
  };
  headers: never;
  pathItems: never;
}